Move nav menu wrapper into main nav menu
